refactor: TodoList画面の機能を一つのディレクトリにまとめた
